refactor(web): update comps path
