+++ packages/web/src/app/(authenticated)/settings/components/DiscordButton.tsx
